Fix instanceof to support Class<> targets in objects
Summary:
The implementation of the `instanceof` test currently matches on the type of the RHS and expects to see a `ClassT (InstanceT)`. If it sees something else within the `ClassT`, it will simply not propagate a lower bound to the refined output type. This is accurate for literal references to classes, but doesn't work in the case of class types arrived at by use of the `Class<>` utility type:

```
class C {}
declare var A: { C: Class<C> };
declare var x: mixed;
if (x instanceof A.C) { ... }
```
In this case, the type of `A.C` in the `instanceof` check is `ClassT (AnnotT (OpenT (lbs=[InstanceT])))`. This ought to behave basically the same as `InstanceT`, but since the matching is direct and structural, it's not, and the result will be `x` being `empty` inside the conditional.

This diff fixes this by relaxing the pattern matching where possible, allowing non-`InstanceT`s inside instanceof checks. We're still not as precise in the result as we would be with an `InstanceT` because an `OpenT` won't have a `instance_id` available for inspection, but at least we won't be unsound.'

This change reveals an unrelated problem in the new env which is addressed in the following diff.

Changelog: [errors] Fixed a bug that caused targets of `instanceof` refinements against `Class<>` annotations to be unsound

Reviewed By: SamChou19815

Differential Revision: D36229307

fbshipit-source-id: 49f1bd3c3ab430d24f96d6a7ff242ec5821099e0