fix invalid export enum exception
Summary:
when parsing `export ...` and we see a `T_ENUM`, we call `Parse.statement_list_item`.

`statement_list_item` only parses a `T_ENUM` as an enum when enums are enabled. if enums are disabled, it falls back to a regular expression statement, and we end up with an invalid export statement and hit a `failwith`.

to fix, we need to check whether enums are enabled; if they're not, it should fall through and expect `export {...`

Changelog: [fix] Fix parser exception when parsing `export enum ...`

Reviewed By: gkz

Differential Revision: D36271858

fbshipit-source-id: 3e7a9b851ddf35645bef9d72aa60c007e17c1677