Tparams in resolved environment
Summary:
This diff builds support for type parameters in the fully resolved environment. It's a little more involved than most of the diffs building the fully resolved environments, due to the way that tparams are represented and substituted.

Part of this diff works pretty similarly to the previous diffs: we (have previously) collected the AST node that defines every type param in a program, and we'll process them in dependency order. When we do process them, we'll figure out the `GenericT` that represents the tvar and store it in the new environment. This is sufficient for type-alias-like uses of tparams, e.g. `function f<X>() { (42: X); }`. When we look up `X` in the body of `f`, we'll find a TypeT (GenericT) for the type and correctly error when we flow NumT into it.

The more complex part is for handling generics within function and class signatures in a way that can be instantiated (explicitly or implicitly). Instantiation works using the `subst` system, which replaces GenericTs by other types. It doesn't look inside OpenTs; it only handles bare types. However, whenever we query the environment, we always get back an OpenT that has been (hopefully, in the resolved env) unified with a definition. This means that, within type_annotation.ml, when looking up a tparam in a signature we need to do something pretty different from what we do when we look up any other type alias by name (including a type parameter that isn't in a signature position, like above). The way this works currently is that tparams are queried from the tparams_map instead of from the environment. So, in order to maintain this behavior and not require deep systemic changes to the system, we need to be able to construct a tparams_map from environment resolution.

To do so we do a few things. First, when we resolve a type parameter declaration, in addition to storing it in its TypeT form in the new env, we also *directly* add the GenericT to a new, separate type map that lives in the loc_env, which contains only type parameter declarations.

Second, when generating name_defs for annotations, we'll now record the location of all type parameters that are in scope for that annotation--for example, in `function f<X>(x: X): X`, we'll record that `X` is in scope for the signature.

Finally, we'll use these recorded type parameters to query the type parameter map in order to generate a tparams_map, which should work identically to the tparams map in the existing system.

Reviewed By: SamChou19815

Differential Revision: D36187844

fbshipit-source-id: 2550ecb017f73b7e8817a5d2a8727fbab16f43c5