Include filtered type in optional chains
Summary:
This diff modifies the typed AST to include a field in OptionalMembers and OptionalCalls recording the type they had assuming that the optional chain succeeded. The standard type field in the typed AST records the *overall* expression type, combining the successful and failed paths. However, for some post-inference analyses it might make sense to only care about the successful case.

The distinction here is apparent in a case like `x?.y.z`. If `x: ?{y: {z: number}}`, then the overall expression has type `void | number`, so that's the appropriate type to give for the expression in the typed AST. However, we might still want to ask, *assuming* that `x` is nonnull, what `x.y` or `x.y.z` look like, since if `x` is null we'll never actually evaluate the .y or .z parts. This is relevant in a following diff where we only care about the nonnull case for detecting sketchy nulls since optional chaining already indicates that null has been considered as a possible output.

We can record this successful type in the typed ast with an additional `'T` entry to the records of optional calls and members, which is populated with the success type when building the typed AST.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D36218627

fbshipit-source-id: 5cdb52e6b884879461ed0ef31f769bc5b88f3ea8