Use post-inference analysis for finding exists checks
Summary: Currently, existence checks are recorded for use in the post-inference sketchies pass during type inference, and depend on the ~cond and other variables passed through the statement.ml process. This isn't easily compatible with the fully resolved environment, which wants to extract AST nodes and typecheck them out of order, without necessarily knowing whether they're in a conditional context or not. This diff moves the add_exists_check call to a separate post-inference pass, which explores the typed_ast and adds appropriate existence checks to the context syntactically based on the position of the expression.

Reviewed By: SamChou19815

Differential Revision: D36218628

fbshipit-source-id: 109591adff954aef0d30e381a322c5822b97908b