Reuse typed parse entities on re-parse
Summary:
This diff removes complexity from add_resolved_requires and add_sig_hash, which
have an update-entity-if-changed behavior, which is complicated by the fact that
re-parsed files have fresh entities, so we need to compare old entities to new
ones.

The trade-off is that re-parsing gets more complicated. Instead of creating
fresh entities, if the file was previously parsed and typed, we copy the
entities over from the old typed parse object.

Reviewed By: mroch

Differential Revision: D35995292

fbshipit-source-id: 9e97331c7677b3061dd34af5a7fecfe590b31cfb