For func params using object destructuring, a default value, and an annotation, type the default value using the annotation
Summary:
Currently, the following is an error:
```
function f({a, b}: {a: number, b?: string}  = {a: 1}) {}
```
Once we eliminate unsealed objects and make the empty object exact, this exposes many errors where people were doing:
```
function f({a}: {a?: number}  = {}) {}
```
Previously the error above would be hidden by the unsafe unsealed object behaviour.

In this change, if there is an annotation, I use that as the type of the default value. We then flow the default expression value to the annotation. This allows the behaviour above, and I think also has nicer error messages.

I have a diff which fixes all the Flow errors this introduces in WWW, but we have to wait until the version of Flow which contains this fix lands first. This is because the fixes means changing the `= {}` to `= {...null}`, as unsealed objects are inexact, but the current lack of errors depends on the lack of safety of unsealed objects.

Changelog: [errors] Improve the type checking of function parameters that use destructuring, have an annotation, and also have a default value

Reviewed By: SamChou19815

Differential Revision: D36362729

fbshipit-source-id: a8ae43019e6df9ba6866ae7f209dedf51d092167