[lti][codemod] Add annotate empty array codemod
Summary:
This diff adds an experimental codemod for annotating empty array in variable declarations. It will attempt to add type annotations in the following cases, as long as it can infer the type of the array literal to be anything other than `Array<empty>` and `Array<any>`:

```
const a = [];
let b = [];
var c = [];
```

This is a major blocker for rolling out LTI, since we will have no context information to type check empty array literals, and we will have to type it as `Array<any>` or `Array<empty>`, which is not ideal.

This is also a major source of error diff between constrained write mode and new-env mode. In the following case in the constrained write mode, we will still flow `Array<number>` to the general type (`Array<OpenT>`) and widen it during the type inference stage:

```
let arr = [];
arr = [3];
```

But we will not do anything at all during type inference in new-env, and it will now error with number is not compatible with elements of empty array. Therefore, we need this codemod to fill in the blank.

Changelog: [internal]

Reviewed By: gkz

Differential Revision: D35783656

fbshipit-source-id: 86bf729b173c48698a33e5db540cb97a682a81b9