[contextual] Scaffolding hint evaluation and usage
Summary:
Changelog: [internal]

## Background

The decomposition of contextual type hint is very lazy. `Hint_Decomp` only describes how types should be decomposed, but the decomposed type hint should only be fully evaluated when we need it. This diff scaffolds the framework code that will evaluate a hint to a type.

## Setup

All of the following setup has one common purpose: if hint decomposition fails, we want it to behave as if there is no hint and have no effect on the current typing context.

The evaluation of the hint is only allowed to proceed if the base hint is fully resolved. This is current done through a potentially expensive runtime check, but the eventually goal is that it will become an invariant once we are able to connect together contextual-typing, fully-resolved env, and implicit instantiation. The fully-resolved requirement is designed to prevent pollution of the typing context with invalid constrains. e.g. if the hint has an unresolved tvar somewhere in its structure, a decomposition might succeed but later might introduces other errors without use_op when the unresolved tvar has additional constrains.

The evaluation of the hint has to be done in a sandboxed environment. This ensures that decomposition produces no type errors even when there are some type incompatibilities. It has to be achieved through a sandbox instead of just disabling type errors, because type errors are sometimes used to guide speculation.

## Testing

Right now it's difficult to test contextual typing in integration tests, because we don't have fully resolved environment yet, so the `is_fully_resolved` test almost always fails. Therefore, I setup unit tests that tests `Type_hint.use_hint` in isolation. It will call `Type_hint.use_hint hint t` and then print the `t` to see whether the type `t` is using the contextually type information as intended.

To achieve that, I need to setup a dummy context, dummy AST, dummy file sig, etc to run the type inference system and the type printing system. The setup is much simpler than the typed_ast test setup, because we only care about types and AST of types here.

Reviewed By: panagosg7

Differential Revision: D36105926

fbshipit-source-id: 10aa290a2c5d57b5246b2182bb05dd5f50633c5f