[contextual] Evaluate hint decompositions for functions
Summary:
This diff implements the first type of hint decomposition: functions. Three variants of function type decompositions are implemented in this diff: decompositions for function parameter, function rest parameters and function return type.

All three implementations follow a common strategy, we create a template function type with all slots being `any`, except for the slot we are interested in, which we fill with an `OpenT`. Then we use `Flow_js.flow` between the hint and the template function type to establish a lower bound for `OpenT`. Note that we have to use `flow` instead of `unify`, since `unify` does not unify rest parameters.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D36255371

fbshipit-source-id: 062201b5482f6301d2e59155bdc424597a65483e