fix order of flow ls output
Summary:
per `readdir(3)`:

> The order in which filenames are read by successive calls to
>  readdir() depends on the filesystem implementation; it is
>  unlikely that the names will be sorted in any fashion.

this leads to unstable output from `flow ls`. `flow ls`'s previous behavior was similar to linux's `ls -U`, the new behavior is similar to `ls`.

to fix, I added a `sort` option to the file walker (`Files.make_next_files`) and use it in `flow ls`, but don't bother during parsing. I want to do it this way instead of sorting the final result in `lsCommand`, because we stream the output.

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D36202196

fbshipit-source-id: 9535ce7b5feea8457edff02e50f8496f801ec070