[new-env] Mark unreachable writes
Summary:
Currently, the code below will cause a crash in the new env.

```
() => {
    return;
    { let i = 0; i = 1; }
//               ^ crash on missing provider
};
```

The provider of `i` in `i = 1` is `i` in `let i = 0`. Since the code is unreachable, we will not add an assigning write, so we will not create a loc-env entry for the location of `i` in `let i = 0`, which leads to the missing provider crash.

Note that the following code doesn't crash, since the assigning write will be added during lexical hoisting.

```
() => {
    return;
    let i = 0; i = 1;
};
```

The solution to this problem is to also mark unreachable writes as `NonAssigningWrite`, since the behavior should be the same other non assigning write due to illegal binding: we don't do anything with them. (e.g. no subtyping against provider).

Name_def_ordering also needs to be updated to ignore all non-assigning write (not ignoring them previously will add some useless nodes in the graph, but now it will cause a topological sort failure if we don't ignore them).

Changelog: [internal]

Reviewed By: jbrown215

Differential Revision: D36389210

fbshipit-source-id: 6c735dcf15b29841361553294ed85af6bb85ffce