[recheck] Fix crash on empty string import during recheck
Summary:
Changelog: [fix] Do not crash during incremental recheck when an empty string import (`e.g. import ''`) is added.

During incremental recheck, we will compute hash to see what has changed. However, for import with empty string, updating the hash with `""` is a no-op, so the hash is not updated and we are in a false conclusion that `resolved_modules` have not changed. This causes a discrepancy between `require_set` and `resolved_modules`, and leads to a crash. This problem first appears in Flow 0.175.

This diff updates the hash computation by using the string length as well. It will create a distinct hash when import of `""` is added, so we have restored the consistency.

Reviewed By: samwgoldman

Differential Revision: D36426652

fbshipit-source-id: b3a9f4d336a6cacda9a774c5ed13c5db38797e54