[contextual] Replace `Decomp_JsxChildren` with combinations of other decompositions
Summary:
After refactors in previous few diffs, it's now clear that `Decomp_JsxChildren` can be replaced by combinations of other decompositions.

Right now, it's only passed into `collapse_children` when we are desugaring jsx in statement.ml. It can be simply replaced by `(Decomp_ObjProp "children")`. The more interesting change is in `collapse_children` when we try to use the hint.

React JSX has some complicated logic of interpreting the children array. In react_kit, [we consider six possible cases](https://github.com/facebook/flow/blob/e1e0af3d1ad92b1f76294fc7789d0dacc6e424d1/src/typing/react_kit.ml#L435-L505): [0, 1, or multiple children arg] x [some or none spread children]. The analysis in the previous diff already shows that we don't need to worry about the presence of spread here, since it's unsupported syntax. (This doesn't mean the code in react_kit can be removed, since it's still valid syntax in a normal `React.createElement` call). Thus, we only need to consider the cases on the number of children nodes.

When there are 1 child, the child will be directly be passed as the value of prop `children`, but when there are more than 1 children, they will be passed in as an array. Therefore, when the child count is 1, the hint `collapse_children` becomes the hint on the child; when the count is more than 1, the hint needs to be decomposed using array element decomposition. When there are 0 children, we don't need to worry about it, since there is nowhere to pass the hint.

This change should have no effect on the current system, since none of the hints are used in any meaningful way.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D36331099

fbshipit-source-id: f024e4c92b1f11e4a951412455e8c9a0926e7020