Heap refinements and instanceof/sentinel refinements in resolved env
Summary:
This diff builds support for refinements in the fully resolved environment. This adds two cases: environment entries for heap refinements, and environment entries for values used in computing instanceof and sentinel refinements.

Heap refinements are introduced into the name_def data structure whenever we see a member or optional member expression that corresponds to an `AssignmentWrite` from the name resolver. We store them as `ChainExpression`s, so named because when we introduce them into the fully resolved environment, the type that we store is the "successful" case of the expression if the expression is an optional chain--that is, if `x: ?{ y: number }` and we create a heap refinement for `x?.y`, the type we store in the environment for this refinement is `number`, not `void | number`. This corresponds to the current behavior of the env in Flow.

One caveat where this approach fails, and which is TODO work, is that in statement.ml we're passing through a `~cond` parameter which is used to decide between `GetPropT` and `TestPropT` for lookups. If a lookup is a heap refinement, we don't currently have a ~cond parameter in the fully resolved environment to pass through, so we'll always treat it as a `GetPropT`. We'll need to do some contextual work here to fix this.

The other aspect of this diff is that some refinements require arbitrary expressions to have their types in the environment for use in computing refinements, namely `instanceof` and sentinel refinements. We introduce a new kind of write for these to the name resolver, a `RefinementWrite`, and in the name def analysis we say that any expression that corresponds to a `RefinementWrite` is the name def for itself. These cases we just call `expression` on, rather than `optional_chain`.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D36354968

fbshipit-source-id: cc1566e3fb9eb5cd1480853c6fd9019770110204