Add code action for inserting local inference annots on func params
Summary:
This diff adds support for an autofix to insert missing annotations on function parameters. Notably, we do not use the types stored in the typed-ast because that inference will not be available to use in the future. Instead, we use types that we manually store at arg -> param flows and use the ty normalizer to convert them into annotations.

If there is no call of the function then we have no information to supply the autofix, so we do not display it.

Changelog: [ide] DEPLOYER: Should this be announced? Added a code action to auto-annotate parameters of functions that have missing-local-annot errors. These errors are gated behind the experimental.enforce_local_annotations flag.

Reviewed By: panagosg7

Differential Revision: D34799544

fbshipit-source-id: 660e578065198e205cb1aa4a69c77bccd14cdb64