Fix instanceof for member expressions in new env
Summary:
There were a couple issues with instanceof in the new env.

First, if an instanceof check like `x instanceof A.B` appeared, the env builder would create both a write entry for A.B, and set its Val.t to point to itself. Then, when we typecheck this, we would call Refinement.get on `A.B`, see that it has an entry in the environment, and return the `OpenT` inhabiting that entry *before we ever did any real typechecking*. So the result is that we'd never actually get a type for `A.B`. This is fixed by making instanceof behave more like sentinels: rather than giving the RHS a val pointing at itself, we just give it a write_entry, and when we use it in the new_env we just directly query the write rather than reading.

The second issue was that we only recorded the write type for instanceof in non-conditional contexts. This diff fixes that oversight by adding a call to Env.record_if_necessary in the appropriate place.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D36230229

fbshipit-source-id: dae2f6e5a540a4b1e36b1e2ab848c1af3d206cd9