remove fail:true from Parsing_service_js
Summary:
`Parser_flow` with `fail:true` does exactly the same amount of work as `fail:false`... it's not a short circuit (we should kill it).

other than that, `Parsing_service_js.do_parse` with `fail:true` skips creating a file sig on an invalid file, which is pretty trivial. it's not worth the complexity.

now all parse errors come through the same path

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D36535260

fbshipit-source-id: aa1919b058be411c920c4ca74ea9093b733a8313