Add reverse dependencies when loading from saved state
Summary:
I forgot to do this in D36488274. It passed tests because I tested under a flag,
which was exclusive with testing saved state.

This diff adds another runtests mode which is incremental rev deps + saved
state, which had failures. This diff fixes.

In needed to change the `prepare_write_sklist` API a little for this. Before, we
passed the file address, but when we load from saved state, we don't have the
file address until we are in the write function. This change adds a bit of noise
to the diff, but the changes are mechanical.

Reviewed By: jbrown215

Differential Revision: D36504793

fbshipit-source-id: d95939f40599031e49911f48a08403efd738da93