Use incremental reverse dependencies during recheck
Summary:
In this diff we use the incrementally maintained reverse dependencies to find
direct dependent files as part of recheck.

When the feature flag is enabled, we use a new method to collect the direct
dependents of changed modules. This method simply iterates over the dependent
set and adds the file keys to a set.

Because the new method is fast, we no longer need the cache. When the feature
flag is enabled, we don't bother caching direct dependents.

Reviewed By: jbrown215

Differential Revision: D36488366

fbshipit-source-id: 370c5964151651331b2a64b2f930eebf4fe4a4bf