[new-env] Perform `super()` call in derived constructor checks in name resolver in a separate pass
Summary:
This diff aims to move one special case handling of `this` and `super` in the old environment to the new environment: the `no-this-before-super` check. It prepares a later diff that completely models `this` and `super` inside name resolver.

As a recap, D9369849 (https://github.com/facebook/flow/commit/6d377c4e9637acc514e67b6d68705d8edce2b3a8) is the diff that implements the check in the current form. Inside a derived constructor, it will initially set `this` and `super` to be undeclared, then a `super()` call will move it into declared and initialized state. All of this happens inside the old environment. It's pretty hard to reason about, since the logic is spread across `class_sig.ml` and `statement.ml`.

In this diff, I reimplement the check as a separate pass over a class in name_resolver. Unlike the rest of the name resolver when it encounters some binding errors that will only error in the new-env mode, this check will unconditionally error. This allows us to completely remove the corresponding code for the old env.

It's important to know that the analysis must be control flow aware. e.g. we shouldn't error in the following case:

```
constructor() {
  if (bool) super() else super()
  super.x
}
```

Therefore, it limits our choice of the core engine to be one of `ssa_builder` and `name_resolver`. It will be quite messy to use `name_resolver` to implement this, because we will then need to inherit and extend `name_resolver` inside `name_resolver`. In addition, this will also breaks an assumption of `name_resolver` that binding cannot be undeclared in one branch but declared in another branch. Hence, `ssa_builder` is the only class that makes sense to extend.

The implementation will need to detect two kinds of errors:
1. using `this` and `super` before calling `super()`: we can perform a post-run analysis: record all locs of this and super reads, and in the end check whether their writes contain uninitialized.
2. calling `super` twice: we record a read at the `super()` call location before writing. In the end, we will error if the read reads all initialized writes.

Changelog: [internal]

Reviewed By: mvitousek

Differential Revision: D36367946

fbshipit-source-id: 0bdea59cb29c260b97f7d2a9c9e0174f5699cf78