[parser] update expected parse errors
Summary:
we test the flow parser against the esprima test suite in src/parser/test/esprima/.

we currently don't handle expected parse errors -- if `foo.js` is expected to have a parse error, the error is in `foo.failure.json`, and we currently skip all of those tests.

the next diff will add support for these tests. this diff fixes the differences so tests don't break on the next diff.

we use `.diff` files so that we don't edit the upstream tests, making it easier to tell where we diverge and future merges easier. for cases where esprima errors and we don't (seems like bugs on our side!), use `.skip` files to continue to skip those tests for now.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D36584915

fbshipit-source-id: 8877a83fb8b3bf0df9a71798489c491b889f3802