[parser] fix crash after unexpected backslash
Summary:
attempting to parse the expression `\/` hit a `failwith` and crashed the parser.

now:
```
  "errors":[
    {
      "loc":{"source":null,"start":{"line":1,"column":0},"end":{"line":1,"column":1}},
      "message":"Unexpected token ILLEGAL"
    },
    {
      "loc":{"source":null,"start":{"line":1,"column":0},"end":{"line":1,"column":1}},
      "message":"Unexpected token `\\`, expected a regular expression"
    },
    {
      "loc":{"source":null,"start":{"line":2,"column":0},"end":{"line":2,"column":0}},
      "message":"Unexpected end of input"
    }
  ]
```

I haven't looked into why we encounter the `\`, error about it, then somehow look ahead at the `/` and say "that must be a regexp" but haven't consumed the `\` yet so it fatals. an additional (future) fix would be to consume the bad `\` so that the `regexp` function parses the `/`. it already knows how to handle an unclosed regex.

Changelog: [parser] Fix crash when encountering an unexpected `\/`. This could have broken IDE integrations while typing.

Reviewed By: jbrown215

Differential Revision: D36631536

fbshipit-source-id: b5c644b7f538b39630963d45bf2bd53c1760c618