[new-env][RFC] Precisely model implicitly defined `this` and `super` in `name_resolver`
Summary:
Unlike the old env, when we are resolving names in the new env, we only use the loc of the name and do not rely on the current scope tracked by the old env. This works extremely well for all kinds of ordinary names, since all of them have an explicit def loc in the source that can be used as a key. This kind of model breaks down for `this` and `super`, since they don't have a clear place of definition. As a result, they are currently not tracked in name resolver, and cause all the refinements on `this` and `super` to be untracked.

Some hacks can be used to overcome the problem, but they are not very clean:

1. Model `this` and `super` as a module scoped variable, and delegate to the old env for types.
    This actually kinda works, but it still leaves a lot of code that depends on scope being tracked by env for this and super.
2. In the parser, at each function, give this and super a unique location for definition.
    This is very unprincipled and leaks type checking behavior into parser.

If we look at the problem at a higher level, we will see that we are unnecessarily constrain ourselves with only location as the key in `Loc_env`. With this constraint in place, we cannot give two different writes the same location, which is what we hope for `this` and `super`: using the function loc as the def loc.

Instead, more generally, the loc env should be keyed by a write, and we can represent the write uniquely, including `this` and `super` as: `(def_loc_type, loc)`, where `def_loc_type` can be `OrdinaryNameLoc | ThisLoc | SuperLoc`. This setup gives us different namespaces of writes, so we now have the full expressive power to model writes. Changes in `env.ml, new_env.ml` are mostly for dealing with the new parameter `def_loc_type`.

Another big part of this diff is to track the reads and writes of `this` precisely in `name_resolver`. (`statement.ml` also needs to be updated, but it's more mechanical.) In summary, the following important changes are made:

- Reads of `this` and `super` are modeled as reads of ordinary name
- Writes of `this` and `super`
  - Have their special Val: `Val.this`, `Val.super`, so that the writes can be distinguished from ordinary name writes in new-env later.
  - Writes to `this` and `super` will be recorded to their own `write_entries`
  - Happens at each function, conditionally:
    - No writes of `this` and `super` in arrow functions, which correctly models the behavior that they bind the entry in upper scopes. As a result, havoc logic needs to be updated to not invalidate this and super vals.
    - For most functions, their def_loc are both the function loc.
    - For class methods, their def_loc are the method names (to reuse the loc already tracked in class sig).
  - Also happens at class properties, since they can use `this` and `super` and are modeled as functions as well in `func_sig`. The def loc is chosen to be the loc of the entire property.

This approach is not without its drawbacks. The biggest issue is that the implementation detail of choices of def loc for `this` and `super` now leaks into type inference stage. However, I think the benefit of accurate modeling out-weights the leaky abstraction here. In addition, it can remove a lot of coupling between syntactic and semantics analysis in the new env. After this change, the only remaining piece is `arguments`, which can be modeled using the same infrastructure introduced in this diff.

Changelog: [internal]

Reviewed By: mvitousek

Differential Revision: D35165546

fbshipit-source-id: 6979c2c6aec5ca21cf73ae4b9846db355d06edbe