[parser] fix crash on "
Summary:
the parser currently crashes on the input `"`.

the parser recovers from the unexpected EOF and returns an empty string literal -- value = `""`, raw = `"\""`.

when we parse directives, we try to remove the string's quotes, but there's only one, so the `String.sub` raises about an invalid argument (length = -1).

Changelog: [parser] Fix crash when the input is a lone quote

Reviewed By: gkz

Differential Revision: D36655654

fbshipit-source-id: dcff37409e83ab9175d31b87639a79cfbfde07e5