[new_website] Underline error ranges
Summary:
Changelog: [internal]

This diff adds underlines to ranges that are covered by errors. It works in the following way:

For error `e` that covers column i to j, we produce nodes of the following form that will be placed directly after all token nodes in a line:

```
<div>
  <span>{i spaces}</span>
  <span style={underline_style}>{j-i spacces}</span>
</div>
```

The first span is used to left padding until we reached the start of the error range. The padding's width should match the text width exactly, since the code font is monospaced. The second span is used to create actual red error underlines.

I think this gives enough visual information for the user on where is the error. Given that in most code blocks we only have one error, we probably don't really need to replicate the old website's behavior of hover to see the error, instead of displaying all errors below the code block.

Reviewed By: gkz

Differential Revision: D36581351

fbshipit-source-id: 35c921722b543c703330377892d93a332ccd17ac