[contextual] Move hint type definition and constructors into env_builder
Summary:
Move it so that name_def can construct them.

This diff also makes the type of hint polymorphic. The base `Hint_t` can attach any types. In name_def, it will be expressions or annotations, and during type checking in fully resolved environment, that slot will be replaced with types.

Changelog: [internal]

Reviewed By: mvitousek

Differential Revision: D36558205

fbshipit-source-id: a1e8057b0856396c3e602f0688a4d1f9eb130f21