[contextual] Remove method decomposition ops on types
Summary:
I plan to move the decomposition ops into env_builder, so that it can produce the decomposition chain to provide enough information for `Contextual` binding. To achieve that, we have to break the dependency on types.

For `Decomp_MethodName`, all we need is the prop name, and for `Decomp_MethodElem` we can assume the element type is string for now. None of this would matter in the long term, since they are likely to be removed.

Reviewed By: mvitousek

Differential Revision: D36558206

fbshipit-source-id: 1dbb68f8314c30b2eed4f787025cb11d9550dcd6