[env_builder] Use contextual hint in fully resolved environment behind a flag
Summary:
This diff connects together contextual typing and fully resolved environment, with the hint we computed in the last diff passed into the contextual variant.

This change added 1 more failure and 2 more crash for `./runtests.sh -cx` when the integration is unconditionally enabled. It will be the responsibilities of future diffs to reduce the error diff and crashes.

Despite there are a still a lot of stuff not working under fully resolved environment and contextual typing, we are able to see the first limited success in the added test case. Although it's successful type checking cannot be completely attributed to contextual typing, since the old inference algorithm is still adding all the constraints, we can see in the verbose trace that the type of `a` in `(a) => a` is correctly assigned to be `StrT`.

Changelog: [internal]

Reviewed By: mvitousek

Differential Revision: D36559255

fbshipit-source-id: 976d0757be3f58f2337099d1b0dbcf330378d01f