fix trailing newline when reading from stdin
Summary:
the way we read from stdin for various commands like `flow ast`, `get-def`, `check-contents`, etc. appends a newline at the end of stdin, even if one didn't exist.

so `echo 'foo' | flow ast` and `echo -n 'foo' | flow ast` look the same, but in certain edge cases they may not be.

the bug is that we used `input_line` in a loop, which returns each line without the newline, until you hit EOF at the beginning of a line. so we can't tell whether the last line has a newline or not. plus, it's inefficient to read one line at a time.

using the existing `read_all` reads 4KB chunks instead, and doesn't have to guess whether there was a newline or not.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D36655220

fbshipit-source-id: 1efd957e3a09e5b655a7514363b2385ac1558bfb