Break dependency on old env for exhaustiveness checking
Summary:
This diff builds exhaustiveness checking support into the new env. In the classic env, exhaustiveness checking was handled via a special internal variable `maybe_exhaustively_checked`, which was initialized by switch statements and populated with whatever was left over from the discriminant after all cases that returned. Since this was a variable, branching control flow merged it in the typical ways, which allows us to overall detect exhaustiveness of complex control flow. At the end of a function scope, this variable was flowed into `FunImplicitVoidReturnT`; if the variable has *any* lower bounds, this will flow void into the function return type; but if the switches exhaustively test and return for each possible case, the tvar will have no lower bounds and the void ~> return flow won't fire.

We can't directly represent `maybe_exhaustively_checked` in the new env, but we can use the name resolver to model it. Just as before, we model it as a `let` bound variable that is assigned to by `switch` statements that are not definitely exhaustive. At the end of function scopes, we inspect the writes to this variable and record both 1. the location of any switch statement that could affect exhaustiveness and 2. whether the variable was definitely initialized. For example,
```
function f(x: E) {
  switch (x) { case E.A: }
  switch (x) { case E.A: case E.B: return }
}
```
Here, the first switch statement will not appear in the writes to `maybe_exhaustively_checked` because it is shadowed by the second switch statement. And
```function g(x: E) {
  if (cond) {
    switch (x) { case E.A: case E.B: return }
  } else { }
}
```
in this case, `maybe_exhaustively_checked` is not definitely initialized.

With this information stored in the context, we and with the change in D36359867 (https://github.com/facebook/flow/commit/bab140394a46749156646090da31850ab344c60e) to record exhaustiveness outputs in the AST, we can then finally reconstruct the behavior of exhaustiveness checking in the new env. When we are typechecking a function, after we've checked the body, we can look up the locations of switch statements that affect the final state of `maybe_exhaustively_checked` from the context, and traverse the AST to collect their exhaustive outputs. We join those into a tvar, along with `void` if `maybe_exhaustively_checked` was not definitely initialized, and flow the result into `FunImplicitVoidReturnT` as before.

This is necessary for the fully resolved environment, but it also runs in the new env.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D36459252

fbshipit-source-id: 0dc42a5d1d353a8852524790f63ca4dd0940f8dd